Add useCountdown hook to Cart page
